Committed all files this time V2

+++ src/utils.c
@@ -1,9 +1,10 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <math.h>
 #include "utils.h"
 
 float sigmoid(float output) {
-    return 1 / (1 + exp(-output));
+    return 1 / (1 + expf(-output));
 }
 
 float rand_init() {
